+++ app/src/main/java/com/aek/yagoubi/sac/ClientActivity.java
@@ -54,7 +54,7 @@ public class ClientActivity extends AppCompatActivity {
             while (res.moveToNext()) {
                 sacs.add(new Sac(res.getInt(0), res.getInt(1), res.getString(2),
                         res.getString(3), res.getString(4), res.getInt(5),
-                        res.getDouble(6), res.getInt(7)));
+                        res.getFloat(6), res.getInt(7)));
             }
 
 
@@ -134,8 +134,6 @@ public class ClientActivity extends AppCompatActivity {
                 public void onClick(View v) {
                     Intent intent1 = new Intent(ClientActivity.this, AjouterSac.class);
                     intent1.putExtra("client", client);
-
-
                     startActivity(intent1);
                 }
             });
@@ -153,6 +151,7 @@ public class ClientActivity extends AppCompatActivity {
                     boolean b = database.updateClientInformation(id, name, Tele);
 
                     if (b) {
+                        //refresh
                         //refresh
                         if (android.os.Build.VERSION.SDK_INT >= 11) {
 
@@ -180,14 +179,14 @@ public class ClientActivity extends AppCompatActivity {
     protected void onResume() {
         AjouterClient database = new AjouterClient(this);
 
+
         ArrayList<Sac> sacs = new ArrayList<>();
         Cursor res = database.getAllArticlesById(client.getId());
-
 
         while (res.moveToNext()) {
             sacs.add(new Sac(res.getInt(0), res.getInt(1), res.getString(2),
                     res.getString(3), res.getString(4), res.getInt(5),
-                    res.getDouble(6), res.getInt(7)));
+                    res.getFloat(6), res.getInt(7)));
         }
